+++ scripts/tatoeba/README.md
@@ -19,13 +19,13 @@ git clone git@github.com:Helsinki-NLP/Tatoeba-Challenge.git
 
 To convert a few models, call the conversion script from command line:
 ```bash
-python src/transformers/convert_marian_tatoeba.py --models heb-eng eng-heb --save_dir converted
+python src/transformers/convert_marian_tatoeba_to_pytorch.py --models heb-eng eng-heb --save_dir converted
 ```
 
 To convert lots of models you can pass your list of Tatoeba model names to `resolver.convert_models` in a python client or script.
 
 ```python
-from transfomers.convert_marian_tatoeba import TatoebaConverter
+from transformers.convert_marian_tatoeba_to_pytorch import TatoebaConverter
 resolver = TatoebaConverter(save_dir='converted')
 resolver.convert_models(['heb-eng', 'eng-heb'])
 ```
